backend/pyproject.toml
[project]
name = "hypernode-backend"
version = "0.1.0"
description = "Backend agent workflow for the Project 2030 prototype."
requires-python = ">=3.13,<3.14"
dependencies = [
  "fastapi[standard]>=0.110,<1.0",
  "google-cloud-discoveryengine>=0.13,<1.0",
  "google-cloud-firestore>=2.16,<3.0",
  "pydantic>=2.7,<3.0",
  "python-dotenv>=1.0,<2.0",
  "google-genai>=1.46,<2.0",
  "genkit; sys_platform != 'win32'",
  "genkit-plugin-google-genai; sys_platform != 'win32'",
  "google-adk>=1.14.1",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
  "pytest>=8.3,<9.0",
  "pytest-asyncio>=0.25,<1.0",
]
